refactor: 구입금액 타입 int > long으로 변경
이유: 금액이 int 값을 초과할수도있다. 총 상금이 int값을 초과 할 수 있어 혹시 모를 충돌방지.

+++ src/main/java/lotto/controller/LottoController.java
@@ -10,12 +10,10 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class LottoController {
-    private int purchaseAmount;
+    private long purchaseAmount;
     private final List<Lotto> lottoTickets = new ArrayList<>();
 
     public void setWinningNumbers() {
-
-        //당첨 번호, 보너스 번호 입력 단계
         List<Integer> numbersInput = InputView.inputWinningNumbers();
         int numberInput = InputView.inputBonusNumber(numbersInput);
         WinningNumbers winningNumbers = WinningNumbers.getInstance(numbersInput, numberInput);
@@ -24,7 +22,7 @@ public class LottoController {
 
     public void purchase() {
         purchaseAmount = InputView.inputPurchaseAmount();
-        int tickets = purchaseAmount / 1000;
+        long tickets = purchaseAmount / 1000;
         LottoNumberGenerator lottoNumberGenerator = new LottoNumberGenerator();
 
         for (int i = 0; i < tickets; i++) {
